+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallsPage/callsTableColumns.tsx
@@ -5,11 +5,8 @@
 
 import {
   GridColDef,
-  GridColumnGroup,
   GridColumnGroupingModel,
-  GridColumnNode,
   GridRenderCellParams,
-  GridValidRowModel,
 } from '@mui/x-data-grid-pro';
 import {Tooltip} from '@wandb/weave/components/Tooltip';
 import {UserLink} from '@wandb/weave/components/UserLink';
@@ -18,13 +15,7 @@ import React, {useCallback, useEffect, useMemo, useState} from 'react';
 import {monthRoundedTime} from '../../../../../../common/util/time';
 import {isWeaveObjectRef, parseRef} from '../../../../../../react';
 import {makeRefCall} from '../../../../../../util/refs';
-import {ErrorBoundary} from '../../../../../ErrorBoundary';
 import {Timestamp} from '../../../../../Timestamp';
-import {CellValue} from '../../../Browse2/CellValue';
-import {CollapseHeader} from '../../../Browse2/CollapseGroupHeader';
-import {ExpandHeader} from '../../../Browse2/ExpandHeader';
-import {NotApplicable} from '../../../Browse2/NotApplicable';
-import {SmallRef} from '../../../Browse2/SmallRef';
 import {Reactions} from '../../feedback/Reactions';
 import {
   getTokensAndCostFromUsage,
@@ -32,6 +23,8 @@ import {
 } from '../CallPage/TraceUsageStats';
 import {CallLink} from '../common/Links';
 import {StatusChip} from '../common/StatusChip';
+import {buildDynamicColumns} from '../common/tabularListViews/columnBuilder';
+import {allOperators} from '../common/tabularListViews/operators';
 import {isRef} from '../common/util';
 import {TraceCallSchema} from '../wfReactInterface/traceServerClientTypes';
 import {
@@ -39,16 +32,8 @@ import {
   traceCallLatencyS,
   traceCallStatusCode,
 } from '../wfReactInterface/tsDataModelHooks';
-import {
-  EXPANDED_REF_REF_KEY,
-  EXPANDED_REF_VAL_KEY,
-  ExpandedRefWithValue,
-  isExpandedRefWithValue,
-  isTableRef,
-} from '../wfReactInterface/tsDataModelHooksCallRefExpansion';
 import {opVersionRefOpName} from '../wfReactInterface/utilities';
 import {OpVersionIndexText} from './CallsTable';
-import {buildTree} from './callsTableBuildTree';
 import {
   insertPath,
   isDynamicCallColumn,
@@ -56,7 +41,6 @@ import {
   stringToPath,
 } from './callsTableColumnsUtil';
 import {WFHighLevelCallFilter} from './callsTableFilter';
-import {allOperators} from './callsTableQuery';
 
 const HIDDEN_DYNAMIC_COLUMN_PREFIXES = ['summary.usage'];
 
@@ -498,158 +482,4 @@ const refIsExpandable = (ref: string): boolean => {
     );
   }
   return false;
-};
-
-export type ExpandedRefWithValueAsTableRef = ExpandedRefWithValue<string>;
-
-const isExpandedRefWithValueAsTableRef = (
-  ref: any
-): ref is ExpandedRefWithValueAsTableRef => {
-  if (!isExpandedRefWithValue(ref)) {
-    return false;
-  }
-  return isTableRef(ref[EXPANDED_REF_VAL_KEY]);
-};
-
-export const buildDynamicColumns = <T extends GridValidRowModel>(
-  filteredDynamicColumnNames: string[],
-  valueForKey: (row: T, key: string) => any,
-  columnIsExpanded?: (col: string) => boolean,
-  columnCanBeExpanded?: (col: string) => boolean,
-  onCollapse?: (col: string) => void,
-  onExpand?: (col: string) => void,
-  columnIsFilterable?: (col: string) => boolean,
-  columnIsSortable?: (col: string) => boolean
-) => {
-  const cols: Array<GridColDef<T>> = [];
-
-  const tree = buildTree([...filteredDynamicColumnNames]);
-  let groupingModel: GridColumnGroupingModel = tree.children.filter(
-    c => 'groupId' in c
-  ) as GridColumnGroup[];
-
-  const walkGroupingModel = (
-    nodes: GridColumnNode[],
-    fn: (node: GridColumnNode) => GridColumnNode
-  ) => {
-    return nodes.map(node => {
-      node = fn(node);
-      if ('children' in node) {
-        node.children = walkGroupingModel(node.children, fn);
-      }
-      return node;
-    });
-  };
-  const groupIds = new Set<string>();
-  groupingModel = walkGroupingModel(groupingModel, node => {
-    if ('groupId' in node) {
-      const key = node.groupId;
-      groupIds.add(key);
-      if (columnIsExpanded && onCollapse && columnIsExpanded(key)) {
-        node.renderHeaderGroup = () => {
-          return (
-            <CollapseHeader
-              headerName={key.split('.').slice(-1)[0]}
-              field={key}
-              onCollapse={onCollapse}
-            />
-          );
-        };
-      } else if (columnCanBeExpanded && onExpand && columnCanBeExpanded(key)) {
-        node.renderHeaderGroup = () => {
-          return (
-            <ExpandHeader
-              headerName={key.split('.').slice(-1)[0]}
-              field={key}
-              hasExpand
-              onExpand={onExpand}
-            />
-          );
-        };
-      }
-    }
-    return node;
-  }) as GridColumnGroupingModel;
-
-  for (const key of filteredDynamicColumnNames) {
-    const col: GridColDef<T> = {
-      flex: 1,
-      minWidth: 150,
-      field: key,
-      filterable: columnIsFilterable && columnIsFilterable(key),
-      sortable: columnIsSortable && columnIsSortable(key),
-      filterOperators: allOperators,
-      headerName: key,
-      renderHeader: () => {
-        return (
-          <div
-            style={{
-              fontWeight: 600,
-            }}>
-            {key.split('.').slice(-1)[0]}
-          </div>
-        );
-      },
-      valueGetter: cellParams => {
-        const val = valueForKey(cellParams.row, key);
-        if (Array.isArray(val) || typeof val === 'object') {
-          try {
-            return JSON.stringify(val);
-          } catch {
-            return val;
-          }
-        }
-        return val;
-      },
-      renderCell: cellParams => {
-        const val = valueForKey(cellParams.row, key);
-        if (val === undefined) {
-          return <NotApplicable />;
-        }
-        return (
-          <ErrorBoundary>
-            {/* In the future, we may want to move this isExpandedRefWithValueAsTableRef condition
-            into `CellValue`. However, at the moment, `ExpandedRefWithValueAsTableRef` is a
-            Table-specific data structure and we might not want to leak that into the
-            rest of the system*/}
-            {isExpandedRefWithValueAsTableRef(val) ? (
-              <SmallRef objRef={parseRef(val[EXPANDED_REF_REF_KEY])} />
-            ) : (
-              <CellValue value={val} />
-            )}
-          </ErrorBoundary>
-        );
-      },
-    };
-
-    if (groupIds.has(key)) {
-      col.renderHeader = () => {
-        return <></>;
-      };
-    } else if (columnIsExpanded && onCollapse && columnIsExpanded(key)) {
-      col.renderHeader = () => {
-        return (
-          <CollapseHeader
-            headerName={key.split('.').slice(-1)[0]}
-            field={key}
-            onCollapse={onCollapse}
-          />
-        );
-      };
-    } else if (columnCanBeExpanded && onExpand && columnCanBeExpanded(key)) {
-      col.renderHeader = () => {
-        return (
-          <ExpandHeader
-            headerName={key.split('.').slice(-1)[0]}
-            field={key}
-            hasExpand
-            onExpand={onExpand}
-          />
-        );
-      };
-    }
-    cols.push(col);
-  }
-
-  return {cols, groupingModel};
 };
